# backend/docs/swagger.yaml
components:
  schemas:
    data:
      properties:
        data:
          $ref: '#/components/schemas/models.UserResponse'
      type: object
...
    handlers.ChangeRecord:
      properties:
        description:
          type: string
        entity:
          type: string
        id:
          type: string
        timestamp:
          type: string
        type:
          type: string
      type: object
...
    handlers.PublishRequest:
      properties:
        changes:
          items:
            $ref: '#/components/schemas/handlers.ChangeRecord'
          type: array
          uniqueItems: false
        template_name:
          type: string
      required:
      - template_name
      type: object
...
    models.Error:
      description: Error response object
      properties:
        code:
          example: "404"
          type: string
        field:
          type: string
        message:
          type: string
      type: object
    models.ErrorResponse:
      description: Generic API error response
      properties:
        code:
          type: integer
        errors:
          items:
            $ref: '#/components/schemas/models.Error'
          type: array
          uniqueItems: false
        message:
          type: string
        meta:
          $ref: '#/components/schemas/models.Meta'
        status:
          type: string
      type: object
...
    models.Meta:
      description: Meta object
      properties:
        timestamp:
          example: "2025-02-12T18:31:40Z"
          type: string
      type: object
...
    models.UserResponse:
      properties:
        created_at:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
        email:
          type: string
        last_login:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
        locale:
          type: string
        name:
          type: string
        profile_picture:
          type: string
        provider:
          type: string
        provider_id:
          type: string
        shops:
          items:
            $ref: '#/components/schemas/models.UserShopResponse'
          type: array
          uniqueItems: false
        user_id:
          description: 'TODO: add sub fields'
          type: string
      type: object
    models.UserShopResponse:
      properties:
        created_at:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
        custom_domain:
          type: string
        shop_id:
          type: integer
        status:
          type: string
        subdomain:
          type: string
        title:
          type: string
        updated_at:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
      type: object
...
  securitySchemes:
    OAuth2AccessCode:
      flows:
        authorizationCode:
          authorizationUrl: http://127.0.0.1:8080/oauth2/auth
          tokenUrl: http://127.0.0.1:8080/oauth2/token
      in: header
      type: oauth2
...
info:
  description: This is the Naytife API documentation
  title: Naytife API Docs
  version: "1.0"
openapi: 3.1.0
paths:
...
  /shops/{shop_id}/publish:
    post:
      description: Queue a build job for the shop's static site
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.PublishRequest'
        description: Publish request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Build job queued successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Shop not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Failed to queue build job
      security:
      - OAuth2AccessCode: []
      summary: Trigger site build
      tags:
      - publish
  /shops/{shop_id}/publish/history:
    get:
      description: Get the publish history for a shop
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      - description: Limit
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      - description: Offset
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Publish history retrieved
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request
      security:
      - OAuth2AccessCode: []
      summary: Get publish history
      tags:
      - publish
  /shops/{shop_id}/publish/status:
    get:
      description: Get the current status of a publish job
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      - description: Job ID
        in: query
        name: job_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Job status retrieved
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Job not found
      security:
      - OAuth2AccessCode: []
      summary: Get publish status
      tags:
      - publish
...
servers:
- url: http://127.0.0.1:8080/v1